+++ tilings/strategies/fusion/__init__.py
@@ -1,10 +1,13 @@
 from .component import ComponentFusionFactory, ComponentFusionStrategy
 from .constructor import FusionConstructor
+from .disjoint_fusion import DisjointFusionFactory, DisjointFusionStrategy
 from .fusion import FusionFactory, FusionRule, FusionStrategy
 
 __all__ = [
     "ComponentFusionFactory",
     "ComponentFusionStrategy",
+    "DisjointFusionFactory",
+    "DisjointFusionStrategy",
     "FusionFactory",
     "FusionStrategy",
     "FusionRule",